Build a mentions-based recommendation engine
See latexrec and mentionsrec for implementation and the
scripts/astrorecs.py script for an example.

Idea is to list papers cited by your article, and count the number of
times mentioned (these are the primary citations).

Of these primary references, find the set of all secondary references
(papers referred to by primary references, but that are not also
primary references).

For each secondary reference, find all tertiary references that are also
primary references. Count the number of mentions.
Then score the secondary reference by the cosine similarity of their
mention count vectors to those of the origin document to primary
references.